+++ nodejs/awsx/apigateway/index.ts
@@ -14,6 +14,7 @@
 
 export * from "./api";
 export * from "./authorizer";
+export * from "./metrics";
 
 // @pulumi/awsx is a deployment-only module.  If someone tries to capture it, and we fail for some
 // reason we want to give a good message about what the problem likely is.  Note that capturing a