+++ nodejs/awsx/ecs/index.ts
@@ -17,6 +17,7 @@ export * from "./taskDefinition";
 
 export * from "./cluster";
 export * from "./image";
+export * from "./metrics";
 export * from "./service";
 export * from "./ec2Service";
 export * from "./fargateService";